requirements.txt
coverage==7.2.2
Django==4.1.7
djangorestframework==3.14.0
numpy==1.24.2
pandas==1.5.3
pip==23.0.1
pyproj==3.5.0
requests==2.28.1
scipy==1.10.1
